pyproject.toml
[project]
name = "agent-trader"
version = "0.1.0"
description = "Agent-centric quantitative research backend"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "pydantic-settings>=2.8.0,<3.0.0",
    "motor>=3.7.0,<4.0.0",
    "influxdb-client>=1.48.0,<2.0.0",
    "langchain>=1.0.0,<2.0.0",
    "langgraph>=1.0.0,<2.0.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-openai>=1.0.0,<2.0.0",
    "httpx>=0.28.1,<1.0.0",
    "apscheduler>=3.11.0,<4.0.0",
    "structlog>=25.1.0,<26.0.0",
    "orjson>=3.10.15,<4.0.0",
    "baostock>=0.8.9,<1.0.0",
    "tushare>=1.4.0,<2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.25.3,<1.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "ruff>=0.9.7,<1.0.0",
    "mypy>=1.15.0,<2.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0,<2.0.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_trader"]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests for functions and modules",
    "agent_integration: Agent integration and node tests",
    "system_flow: System flow tests",
    "live: External live integration tests",
]

[tool.mypy]
python_version = "3.10"
plugins = []
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
namespace_packages = true
explicit_package_bases = true
packages = ["agent_trader"]
